[dir view] updated 'top operation list' for current active dirent
* added `getDirentOperationList` to utils.js
* improved code
